work on Project

+++ Confetti/lib/Test.rb
@@ -4,50 +4,81 @@ require 'Confetti/lib/Confetti.rb'
 
 module Confetti
 
-TEST_MODE = 1
-TEST_WITH_CLEARCASE = 0
+TEST_MODE = true
 
 #----------------------------------------------------------------------------------------------
 
 class Test < MiniTest::Test
-	attr_reader :path
+	attr_reader :path, :root_vob
 
 	def initialize(name)
-		@path = '.tests/' + Time.now.strftime("%y%m%d-%H%M%S")
-		while File.directory?(path)
-			@path = '.tests/' + Time.now.strftime("%y%m%d-%H%M%S%L")
+		if create_fs?
+			@path = '.tests/' + Time.now.strftime("%y%m%d-%H%M%S")
+			while File.directory?(path)
+				@path = '.tests/' + Time.now.strftime("%y%m%d-%H%M%S%L")
+			end
+			FileUtils.mkdir_p(@path)
+			Bento.unzip(fs_zip, @path)
 		end
-		FileUtils.mkdir_p(@path)
-		Bento.unzip(testzip, @path)
+
+		if create_vob? && TEST_WITH_CLEARCASE
+			@root_vob = ClearCASE::VOB.create('', file: vob_zip)
+		end
 
 		super(name)
 		
 		@@current_test = self
 	end
 
-	def testzip
-		'test.fs.zip'
-	end
-
-	def keep?
-		false
+	def teardown
+		cleanup
 	end
 
 	def cleanup
 		@@current_test = nil
 
 		return if keep?
-		FileUtils.rm_r(@path) rescue ''
+		if create_fs?
+			Confetti::DB.cleanup
+			FileUtils.rm_r(@path) rescue ''
+		end
+		if create_vob?
+			@root_vob.remove!
+		end
 	end
-	
-	def teardown
-		cleanup
+
+	#------------------------------------------------------------------------------------------	
+
+	def keep?
+		false
 	end
+
+	def create_fs?
+		true
+	end
+
+	def create_vob?
+		false
+	end
+
+	def fs_zip
+		'test.fs.zip'
+	end
+
+	def vob_zip
+		'test.vob.zip'
+	end
+
+	#------------------------------------------------------------------------------------------	
 
 	def root
 		@path
 	end
 	
+	def root_vob
+		@root_vob
+	end
+
 	def Test.current
 		@@current_test
 	end
